Start sharing stubs, use psalm stubs for phpstan
