+++ .gitattributes
@@ -25,9 +25,8 @@ package-lock.json export-ignore
 package.json export-ignore
 phpcs.xml export-ignore
 phpunit.xml export-ignore
-psalm.stubs.php export-ignore
 psalm.xml export-ignore
-phpstan.stubs.php export-ignore
+stubs.php export-ignore
 phpstan.neon export-ignore
 readme.md export-ignore
 tests export-ignore
