+++ phpstan.neon
@@ -1,7 +1,8 @@
 parameters:
 	reportUnmatchedIgnoredErrors: false
 	bootstrapFiles:
-		- phpstan.stubs.php
+		- vendor/php-stubs/wordpress-stubs/wordpress-stubs.php
+		- stubs.php
 	excludePaths:
 		- */node_modules/*
 		- */tests/*
@@ -69,3 +70,4 @@ parameters:
 		- '#always exists and is not falsy.#'
 		- '#\@param tag must not be named \$this#'
 		- '#Expected 2 \@param tags, found 4\.#'
+		- '#should always throw an exception or terminate script execution but#'
