+++ psalm.xml
@@ -16,7 +16,7 @@
 	</projectFiles>
 	<stubs>
 		<file name="vendor/php-stubs/wordpress-stubs/wordpress-stubs.php" />
-		<file name="psalm.stubs.php" />
+		<file name="stubs.php" />
 	</stubs>
 	<issueHandlers>
 		<UndefinedClass>
